Lesson-05/src/main/java/hillel/course/spring_annotation/example/entity/Player.java
package hillel.course.spring_annotation.example.entity;

import org.springframework.beans.factory.config.ConfigurableBeanFactory;
import org.springframework.context.annotation.Scope;
import org.springframework.stereotype.Component;

@Component("playerBean")
@Scope(value = ConfigurableBeanFactory.SCOPE_PROTOTYPE)
public class Player {
    private String name;
    private String surname;
    private int age;
    private String position;
    private Team team;
    private int count = 1;

    public Player() {
    }

    public Player(String name, String surname, int age) {
        this.name = name;
        this.surname = surname;
        this.age = age;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getSurname() {
        return surname;
    }

    public void setSurname(String surname) {
        this.surname = surname;
    }

    public int getAge() {
        return age;
    }

    public void setAge(int age) {
        this.age = age;
    }

    public Team getTeam() {
        return team;
    }

    public void setTeam(Team team) {
        this.team = team;
    }

    public String getPosition() {
        return position;
    }

    public void setPosition(String position) {
        this.position = position;
    }

    public int getCount() {
        return count;
    }

    public void printNumber() {
        System.out.println("The current number for playerBean: " + count);
        count++;
        if (count == 11) {
            count = 1;
        }
    }

    @Override
    public String toString() {
        return "Player{" +
                "name='" + name + '\'' +
                ", surname='" + surname + '\'' +
                ", age=" + age +
                ", position='" + position + '\'' +
                ", team=" + team +
                '}';
    }
}
